// generated/orval/fetch/google-health-api/models/nutrientQuantityNutrient.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Required. The nutrient type.
 */
export type NutrientQuantityNutrient = typeof NutrientQuantityNutrient[keyof typeof NutrientQuantityNutrient];


export const NutrientQuantityNutrient = {
  NUTRIENT_UNSPECIFIED: 'NUTRIENT_UNSPECIFIED',
  BIOTIN: 'BIOTIN',
  CAFFEINE: 'CAFFEINE',
  CALCIUM: 'CALCIUM',
  CHLORIDE: 'CHLORIDE',
  CARBOHYDRATES: 'CARBOHYDRATES',
  CHOLESTEROL: 'CHOLESTEROL',
  CHROMIUM: 'CHROMIUM',
  COPPER: 'COPPER',
  DIETARY_FIBER: 'DIETARY_FIBER',
  FOLIC_ACID: 'FOLIC_ACID',
  IODINE: 'IODINE',
  IRON: 'IRON',
  MAGNESIUM: 'MAGNESIUM',
  MANGANESE: 'MANGANESE',
  MOLYBDENUM: 'MOLYBDENUM',
  MONOUNSATURATED_FAT: 'MONOUNSATURATED_FAT',
  NIACIN: 'NIACIN',
  PANTOTHENIC_ACID: 'PANTOTHENIC_ACID',
  PHOSPHORUS: 'PHOSPHORUS',
  POLYUNSATURATED_FAT: 'POLYUNSATURATED_FAT',
  POTASSIUM: 'POTASSIUM',
  PROTEIN: 'PROTEIN',
  RIBOFLAVIN: 'RIBOFLAVIN',
  SATURATED_FAT: 'SATURATED_FAT',
  SELENIUM: 'SELENIUM',
  SODIUM: 'SODIUM',
  SUGAR: 'SUGAR',
  THIAMIN: 'THIAMIN',
  TRANS_FAT: 'TRANS_FAT',
  UNSATURATED_FAT: 'UNSATURATED_FAT',
  VITAMIN_A: 'VITAMIN_A',
  VITAMIN_B12: 'VITAMIN_B12',
  VITAMIN_B6: 'VITAMIN_B6',
  VITAMIN_C: 'VITAMIN_C',
  VITAMIN_D: 'VITAMIN_D',
  VITAMIN_E: 'VITAMIN_E',
  VITAMIN_K: 'VITAMIN_K',
  ZINC: 'ZINC',
  FOLATE: 'FOLATE',
} as const;