// generated/orval/fetch/google-health-api/models/nutrientQuantity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */
import type { NutrientQuantityNutrient } from './nutrientQuantityNutrient';
import type { WeightQuantity } from './weightQuantity';

/**
 * Represents the quantity of a nutrient.
 */
export interface NutrientQuantity {
  /** Required. The nutrient type. */
  nutrient?: NutrientQuantityNutrient;
  /** Required. The quantity of the nutrient, measured in grams. */
  quantity?: WeightQuantity;
}